backend/requirements.txt
fastapi
uvicorn
yt-dlp
openai
python-multipart
ffmpeg-python
python-dotenv
